+++ lib/auth/storage/storage_windows.go
@@ -17,7 +17,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package state
+package storage
 
 import (
 	"context"
